requirements.txt
appdirs==1.4.4
asgiref==3.2.7
boto3==1.14.47
botocore==1.17.47
certifi==2020.4.5.2
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.4
django-allauth==0.42.0
django-ckeditor==5.9.0
django-comments-dab==1.4.0
django-crispy-forms==1.9.0
django-facebook-comments==0.1.5
django-filter==2.3.0
django-heroku==0.3.1
django-import-export==2.3.0
django-js-asset==1.2.2
django-storages==1.9.1
django-tinymce==3.0.2
docutils==0.15.2
et-xmlfile==1.0.1
filelock==3.0.12
gunicorn==20.0.4
idna==2.9
jdcal==1.4.1
jmespath==0.10.0
MarkupPy==1.14
mysql==0.0.2
mysqlclient==1.4.6
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
Pillow==7.1.1
pipenv==2020.8.13
psycopg2==2.8.4
PyMySQL==0.9.3
python-dateutil==2.8.1
python-decouple==3.3
python-git==2018.2.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
urllib3==1.25.9
virtualenv==20.0.31
virtualenv-clone==0.5.4
whitenoise==5.1.0
xlrd==1.2.0
xlwt==1.3.0
